requirements.txt
aenum==3.1.15
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
APScheduler==3.10.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
azure-common==1.1.28
azure-core==1.30.0
azure-data-tables==12.5.0
azure-functions==1.18.0
azure-identity==1.15.0
azure-mgmt-core==1.4.0
azure-mgmt-network==25.2.0
azure-mgmt-resource==23.0.1
azure-mgmt-subscription==3.1.1
azure-storage-blob==12.19.1
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
bs4==0.0.2
cachelib==0.12.0
cachetools==5.3.3
case-conversion==2.1.0
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.1
click==8.1.7
clipboard==0.0.4
colorama==0.4.6
colour==0.1.5
comm==0.2.1
cowsay==6.1
cryptography==42.0.2
cursor==1.3.5
debugpy==1.8.1
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.7
distro==1.9.0
docstring_parser==0.16
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.0
Flask==2.3.3
Flask-APScheduler==1.13.1
fqdn==1.5.1
freezegun==1.4.0
fuzzywuzzy==0.18.0
gender-guesser==0.4.0
gitdb==4.0.11
GitPython==3.1.32
google-api-core==2.18.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-cloud-aiplatform==1.47.0
google-cloud-bigquery==3.20.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==21.2.0
h11==0.14.0
html-text==0.5.2
httpcore==1.0.4
httpx==0.27.0
huepy==1.2.1
humanize==4.8.0
idna==3.4
imageio==2.34.0
imageio-ffmpeg==0.4.9
infi.systray==0.1.12
inflect==7.0.0
ipykernel==6.29.3
ipython==8.22.2
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.22
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.4
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.13.0
jupyter_server_terminals==0.5.2
jupyterlab==4.1.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
lxml==5.1.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
maybe-else==0.2.1
mbstrdecoder==1.1.3
mistune==3.0.2
moviepy==1.0.3
msal==1.26.0
msal-extensions==1.1.0
msoffcrypto-tool==5.3.1
msrest==0.7.1
multidict==6.0.5
nbclient==0.9.0
nbconvert==7.16.2
nbformat==5.9.2
nest-asyncio==1.6.0
notebook==7.1.1
notebook_shim==0.2.4
numpy==1.26.4
O365==2.0.34
oauthlib==3.2.2
office365==0.3.15
Office365-REST-Python-Client==2.5.6
olefile==0.47
openai==1.27.0
openpyxl==3.1.2
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parsedatetime==2.6
parso==0.8.3
pathmagic==0.3.14
pillow==10.2.0
platformdirs==3.11.0
portalocker==2.8.2
prettierfier==1.0.3
proglog==0.1.10
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pyinstrument==4.6.2
pyiotools==0.3.18
PyJWT==2.8.0
pymiscutils==0.3.14
PyPDF2==3.0.1
pyperclip==1.8.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pysubtypes==0.3.18
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==0.19.1
python-json-logger==2.0.7
pytz==2023.3
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
readchar==4.0.5
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9
Send2Trash==1.8.2
setuptools==68.2.2
shapely==2.0.3
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
stringcase==1.2.0
tabulate==0.9.0
terminado==0.18.0
tinycss2==1.2.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typepy==1.3.2
types-python-dateutil==2.8.19.20240106
typing_extensions==4.8.0
tzdata==2024.1
tzlocal==5.2
uri-template==1.3.0
urllib3==1.26.18
virtualenv==20.24.6
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
wheel==0.42.0
XlsxWriter==3.2.0
yarl==1.9.4
